pyproject.toml
[tool.poetry]
name = "jogging_tracker"
version = "0.1.0"
description = "REST API that tracks jogging times of users"
authors = ["Tiago Carreira <tiago.g.carreira+git@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
django = "^3.0.5"
gunicorn = "^20.0.4"
djangorestframework = "^3.11.0"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.8.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.4.4"
black = "^19.10b0"
pylint-django = "^2.0.15"
django-discover-runner = "^1.0"
coverage = "^5.1"
pytest-django = "^3.9.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
